user_module/requirements.txt
pypika==0.48.8
pinject==0.14.1
python-dotenv==0.19.2
PyJWT==2.3.0
